Exclude disk information on iOS builds
