+++ Cargo.toml
@@ -28,7 +28,7 @@ libc = "0.2"
 [target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
 core-foundation-sys = "0.8"
 
-[target.'cfg(any(target_os = "macos", target_os = "ios"))'.build-dependencies]
+[target.'cfg(target_os = "macos")'.build-dependencies]
 cc = "1.0"
 
 [lib]
