+++ src/mac/disk/disk.m
@@ -1,9 +1,4 @@
-#include "TargetConditionals.h"
-#if TARGET_OS_IPHONE || TARGET_IPHONE_SIMULATOR
-    #import <UIKit/UIKit.h>
-#else
-    #import <AppKit/AppKit.h>
-#endif
+#import <AppKit/AppKit.h>
 #import <Foundation/NSFileManager.h>
 
 CFArrayRef macos_get_disks() {
