+++ src/mac/disk/ios.rs
@@ -1,0 +1,12 @@
+//
+// Sysinfo
+//
+// Copyright (c) 2021 Guillaume Gomez
+//
+
+use Disk;
+use sys::ffi;
+
+pub(crate) fn get_disks(_: ffi::DASessionRef) -> Vec<Disk> {
+    Vec::new()
+}
